@@ -19,7 +19,8 @@
     }
   ],
   "require": {
-    "php": ">=5.3.0"
+    "php": ">=5.5.0",
+    "ext-json": "*"
   },
   "require-dev": {
     "phpunit/phpunit": "^6.5"
