@@ -70,7 +70,7 @@ abstract class AbstractCallback implements ConstInterface
      *
      * @var array
      */
-    protected $parameters = array();
+    protected $parameters = [];
 
     /**
      * The actual callback function for the renderer.
@@ -171,13 +171,15 @@ abstract class AbstractCallback implements ConstInterface
      */
     protected function isPropertyNameNormal($propName)
     {
-        static $cache = array();
+        static $cache = [];
 
         if (isset($cache[$propName])) {
             return $cache[$propName];
         }
-        $cache[$propName] = (bool) preg_match("/^[a-zA-Z_\x7f-\xff][a-zA-Z0-9_\x7f-\xff]*$/", $propName);
 
-        return $cache[$propName];
+        // The first regex detects all allowed characters.
+        // For some reason, they also allow BOM characters.
+        return $cache[$propName] = (bool) preg_match("/^[a-zA-Z_\x7f-\xff][a-zA-Z0-9_\x7f-\xff]*$/", $propName) &&
+            !(bool) preg_match("/[\xEF\xBB\xBF]$/", $propName);
     }
 }
