@@ -35,9 +35,13 @@
 namespace Brainworxx\Krexx\Analyse\Callback\Iterate;
 
 use Brainworxx\Krexx\Analyse\Callback\AbstractCallback;
-use Brainworxx\Krexx\Analyse\Model;
 use Brainworxx\Krexx\Analyse\Code\Connectors;
+use Brainworxx\Krexx\Analyse\Comment\Methods;
+use Brainworxx\Krexx\Analyse\Model;
 use Brainworxx\Krexx\Service\Factory\Pool;
+use ReflectionClass;
+use ReflectionException;
+use ReflectionMethod;
 
 /**
  * Getter method analysis methods.
@@ -69,6 +73,20 @@ class ThroughGetter extends AbstractCallback
     const CURRENT_PREFIX = 'currentPrefix';
 
     /**
+     * Stuff we need to escape in a regex.
+     *
+     * @var array
+     */
+    protected $regexEscapeFind = ['.', '/', '(', ')', '<', '>', '$'];
+
+    /**
+     * Stuff the escaped regex stuff.
+     *
+     * @var array
+     */
+    protected $regexEscapeReplace = ['\.', '\/', '\(', '\)', '\<', '\>', '\$'];
+
+    /**
      * {@inheritdoc}
      */
     protected static $eventPrefix = 'Brainworxx\\Krexx\\Analyse\\Callback\\Iterate\\ThroughGetter';
@@ -95,7 +113,7 @@ class ThroughGetter extends AbstractCallback
     public function __construct(Pool $pool)
     {
         parent::__construct($pool);
-        $this->commentAnalysis = $this->pool->createClass('Brainworxx\\Krexx\\Analyse\\Comment\\Methods');
+        $this->commentAnalysis = $this->pool->createClass(Methods::class);
     }
 
     /**
@@ -146,7 +164,7 @@ class ThroughGetter extends AbstractCallback
             ));
 
             /** @var Model $model */
-            $model = $this->pool->createClass('Brainworxx\\Krexx\\Analyse\\Model')
+            $model = $this->pool->createClass(Model::class)
                 ->setName($reflectionMethod->getName())
                 ->addToJson(static::META_METHOD_COMMENT, $comments);
 
@@ -158,13 +176,17 @@ class ThroughGetter extends AbstractCallback
             }
 
             // Get ourselves a possible return value
-            $output .= $this->retrievePropertyValue(
-                $reflectionMethod,
-                $this->dispatchEventWithModel(
-                    __FUNCTION__ . static::EVENT_MARKER_END,
-                    $model
-                )
-            );
+            try {
+                $output .= $this->retrievePropertyValue(
+                    $reflectionMethod,
+                    $this->dispatchEventWithModel(
+                        __FUNCTION__ . static::EVENT_MARKER_END,
+                        $model
+                    )
+                );
+            } catch (ReflectionException $e) {
+                // Do nothing. We ignore this one.
+            }
         }
 
         return $output;
@@ -178,10 +200,12 @@ class ThroughGetter extends AbstractCallback
      * @param Model $model
      *   The model so far.
      *
+     * @throws \ReflectionException
+     *
      * @return string
      *   The rendered markup.
      */
-    protected function retrievePropertyValue(\ReflectionMethod $reflectionMethod, Model $model)
+    protected function retrievePropertyValue(ReflectionMethod $reflectionMethod, Model $model)
     {
         /** @var \Brainworxx\Krexx\Service\Reflection\ReflectionClass $reflectionClass */
         $reflectionClass = $this->parameters[static::PARAM_REF];
@@ -204,12 +228,12 @@ class ThroughGetter extends AbstractCallback
         // Give the plugins the opportunity to do something with the value, or
         // try to resolve it,  if nothing was found.
         // We also add the stuff, that we were able to do so far.
-        $this->parameters[static::PARAM_ADDITIONAL] = array(
+        $this->parameters[static::PARAM_ADDITIONAL] = [
             'nothingFound' => $nothingFound,
             'value' => $value,
             'refProperty' => $refProp,
             'refMethod' => $reflectionMethod
-        );
+        ];
         $this->dispatchEventWithModel(__FUNCTION__ . '::resolving', $model);
 
         if ($this->parameters[static::PARAM_ADDITIONAL]['nothingFound'] === true) {
@@ -240,11 +264,13 @@ class ThroughGetter extends AbstractCallback
      *   The reflection ot the method of which we want to coax the result from
      *   the class or sourcecode.
      *
+     * @throws \ReflectionException
+     *
      * @return \ReflectionProperty|null
      *   Either the reflection of a possibly associated Property, or null to
      *   indicate that we have found nothing.
      */
-    protected function getReflectionProperty(\ReflectionClass $classReflection, \ReflectionMethod $reflectionMethod)
+    protected function getReflectionProperty(ReflectionClass $classReflection, ReflectionMethod $reflectionMethod)
     {
         // We may be facing different writing styles.
         // The property we want from getMyProperty() should be named myProperty,
@@ -326,7 +352,7 @@ class ThroughGetter extends AbstractCallback
      * @return string
      *   The first impression of the property name.
      */
-    protected function preparePropertyName(\ReflectionMethod $reflectionMethod)
+    protected function preparePropertyName(ReflectionMethod $reflectionMethod)
     {
         $currentPrefix = $this->parameters[static::CURRENT_PREFIX];
 
@@ -356,11 +382,13 @@ class ThroughGetter extends AbstractCallback
      *   The reflection ot the method of which we want to coax the result from
      *   the class or sourcecode.
      *
+     * @throws \ReflectionException
+     *
      * @return \ReflectionProperty|null
      *   Either the reflection of a possibly associated Property, or null to
      *   indicate that we have found nothing.
      */
-    protected function getReflectionPropertyDeep(\ReflectionClass $classReflection, \ReflectionMethod $reflectionMethod)
+    protected function getReflectionPropertyDeep(ReflectionClass $classReflection, ReflectionMethod $reflectionMethod)
     {
         // Read the sourcecode into a string.
         $sourcecode = $this->pool->fileService->readFile(
@@ -372,7 +400,7 @@ class ThroughGetter extends AbstractCallback
         // Execute our search pattern.
         // Right now, we are trying to get to properties that way.
         // Later on, we may also try to parse deeper for stuff.
-        foreach ($this->findIt(array('return $this->', ';'), $sourcecode) as $propertyName) {
+        foreach ($this->findIt(['return $this->', ';'], $sourcecode) as $propertyName) {
             // Check if this is a property and return the first we find.
             $parentClass = $classReflection;
             while ($parentClass !== false) {
@@ -414,7 +442,7 @@ class ThroughGetter extends AbstractCallback
      */
     protected function convertToSnakeCase($string)
     {
-        return strtolower(preg_replace(array('/([a-z\d])([A-Z])/', '/([^_])([A-Z][a-z])/'), '$1_$2', $string));
+        return strtolower(preg_replace(['/([a-z\d])([A-Z])/', '/([^_])([A-Z][a-z])/'], '$1_$2', $string));
     }
 
     /**
@@ -462,8 +490,8 @@ class ThroughGetter extends AbstractCallback
     protected function regexEscaping($string)
     {
         return str_replace(
-            array('.', '/', '(', ')', '<', '>', '$'),
-            array('\.', '\/', '\(', '\)', '\<', '\>', '\$'),
+            $this->regexEscapeFind,
+            $this->regexEscapeReplace,
             $string
         );
     }
