@@ -35,6 +35,8 @@
 namespace Brainworxx\Krexx\Analyse\Comment;
 
 use Brainworxx\Krexx\Service\Factory\Pool;
+use ReflectionClass;
+use Reflector;
 
 /**
  * Abstract class for the comment analysis.
@@ -54,12 +56,12 @@ abstract class AbstractComment
      *
      * @var array
      */
-    protected $inheritdocPattern = array(
+    protected $inheritdocPattern = [
         '{@inheritdoc}',
         '{inheritdoc}',
         '@inheritdoc',
         'inheritdoc'
-    );
+    ];
 
     /**
      * Inject the pool
@@ -83,8 +85,8 @@ abstract class AbstractComment
      *   The prettified comment.
      */
     abstract public function getComment(
-        \Reflector $reflection,
-        \ReflectionClass $reflectionClass = null
+        Reflector $reflection,
+        ReflectionClass $reflectionClass = null
     );
 
     /**
@@ -105,7 +107,7 @@ abstract class AbstractComment
         // We split our comment into single lines and remove the unwanted
         // comment chars with the array_map callback.
         // We skip lines with /** and */
-        $result = array();
+        $result = [];
         foreach (array_slice(explode("\n", $comment), 1, -1) as $commentLine) {
             // Remove comment-chars and trim the whitespace.
             $result[] = trim($commentLine, "* \t\n\r\0\x0B");
