@@ -34,6 +34,11 @@
 
 namespace Brainworxx\Krexx\Analyse\Comment;
 
+use ReflectionClass;
+use ReflectionException;
+use ReflectionMethod;
+use Reflector;
+
 /**
  * We get the comment of a method and try to resolve the inheritdoc stuff.
  *
@@ -63,10 +68,10 @@ class Methods extends AbstractComment
      * @return string
      *   The prettified and escaped comment.
      */
-    public function getComment(\Reflector $reflectionMethod, \ReflectionClass $reflectionClass = null)
+    public function getComment(Reflector $reflectionMethod, ReflectionClass $reflectionClass = null)
     {
         // Do some static caching. The comment will not change during a run.
-        static $cache = array();
+        static $cache = [];
         /** @var \ReflectionMethod $reflectionMethod */
         $this->methodName = $reflectionMethod->getName();
         $cachingKey =  $reflectionMethod->getDeclaringClass()->name . '::' . $this->methodName;
@@ -93,7 +98,7 @@ class Methods extends AbstractComment
      * @return string
      *   The prettified comment.
      */
-    protected function getMethodComment(\ReflectionMethod $reflectionMethod, \ReflectionClass $reflectionClass)
+    protected function getMethodComment(ReflectionMethod $reflectionMethod, ReflectionClass $reflectionClass)
     {
         // Get a first impression.
         $comment = $this->prettifyComment($reflectionMethod->getDocComment());
@@ -120,14 +125,19 @@ class Methods extends AbstractComment
         // Nothing on this level, we need to take a look at the parent.
         /** @var \ReflectionClass $parentReflection */
         $parentReflection = $reflectionClass->getParentClass();
-        if ($parentReflection instanceof \ReflectionClass &&
+        if ($parentReflection instanceof ReflectionClass &&
             $parentReflection->hasMethod($this->methodName) === true
         ) {
             // Going deeper into the rabid hole!
-            $comment = $this->getMethodComment(
-                $parentReflection->getMethod($this->methodName),
-                $parentReflection
-            );
+            try {
+                $comment = $this->getMethodComment(
+                    $parentReflection->getMethod($this->methodName),
+                    $parentReflection
+                );
+            } catch (ReflectionException $e) {
+                // Do nothing.
+                // The feedback is a little bit further down.
+            }
         }
 
         // Still here? Tell the dev that we could not resolve the comment.
@@ -150,34 +160,33 @@ class Methods extends AbstractComment
      * @return string
      *   The comment from one of the trait.
      */
-    protected function getTraitComment($originalComment, \ReflectionClass $reflection)
+    protected function getTraitComment($originalComment, ReflectionClass $reflection)
     {
-        // We need to check if we can get traits here.
-        if (method_exists($reflection, 'getTraits') === true) {
-            // Get the traits from this class.
-            // Now we should have an array with reflections of all
-            // traits in the class we are currently looking at.
-            foreach ($reflection->getTraits() as $trait) {
-                if ($this->checkComment($originalComment) === true) {
-                    // Looks like we've resolved them all.
-                    return $originalComment;
-                }
+        // Get the traits from this class.
+        // Now we should have an array with reflections of all
+        // traits in the class we are currently looking at.
+        foreach ($reflection->getTraits() as $trait) {
+            if ($this->checkComment($originalComment) === true) {
+                // Looks like we've resolved them all.
+                return $originalComment;
+            }
 
-                // We need to look further!
-                if ($trait->hasMethod($this->methodName) === true) {
+            // We need to look further!
+            if ($trait->hasMethod($this->methodName) === true) {
+                try {
                     $traitComment = $this->prettifyComment(
                         $trait->getMethod($this->methodName)->getDocComment()
                     );
                     // Replace it.
                     $originalComment = $this->replaceInheritComment($originalComment, $traitComment);
+                } catch (ReflectionException $e) {
+                    // Do nothing.
+                    // We could not resolve anything.
                 }
             }
-
-            // Return what we could resolve so far.
-            return $originalComment;
         }
 
-        // Wrong PHP version. Traits are not available.
+        // Return what we could resolve so far.
         return $originalComment;
     }
 
@@ -195,13 +204,20 @@ class Methods extends AbstractComment
      * @return string
      *   The comment from one of the interfaces.
      */
-    protected function getInterfaceComment($originalComment, \ReflectionClass $reflectionClass)
+    protected function getInterfaceComment($originalComment, ReflectionClass $reflectionClass)
     {
         foreach ($reflectionClass->getInterfaces() as $interface) {
             if ($interface->hasMethod($this->methodName) === true) {
-                $interfaceComment = $this->prettifyComment($interface->getMethod($this->methodName)->getDocComment());
-                // Replace it.
-                $originalComment = $this->replaceInheritComment($originalComment, $interfaceComment);
+                try {
+                    $interfaceComment = $this->prettifyComment($interface->getMethod($this->methodName)
+                        ->getDocComment());
+                    // Replace it.
+                    $originalComment = $this->replaceInheritComment($originalComment, $interfaceComment);
+                } catch (ReflectionException $e) {
+                    // Failed to retrieve it from the interface.
+                    // Do nothing, and hope for the rst of the code to retrieve
+                    // the comment.
+                }
             }
             if ($this->checkComment($originalComment) === true) {
                 // Looks like we've resolved them all.
