@@ -48,14 +48,14 @@ class TimerController extends AbstractController
      *
      * @var array
      */
-    protected static $timekeeping = array();
+    protected static $timekeeping = [];
 
     /**
      * More timekeeping stuff.
      *
      * @var array
      */
-    protected static $counterCache = array();
+    protected static $counterCache = [];
 
     /**
      * We simply set the pool. We will not register any shutdown stuff.
@@ -104,11 +104,11 @@ class TimerController extends AbstractController
         $this->timerAction('end');
         // And we are done. Feedback to the user.
         $miniBench = $this->miniBenchTo(static::$timekeeping);
-        $this->pool->createClass('Brainworxx\\Krexx\\Controller\\DumpController')
+        $this->pool->createClass(DumpController::class)
             ->dumpAction($miniBench, 'kreXX timer');
         // Reset the timer vars.
-        static::$timekeeping = array();
-        static::$counterCache = array();
+        static::$timekeeping = [];
+        static::$counterCache = [];
 
         return $this;
     }
@@ -146,5 +146,57 @@ class TimerController extends AbstractController
         }
 
         return $result;
+    }
+
+    /**
+     * Return the current URL.
+     *
+     * @see http://stackoverflow.com/questions/6768793/get-the-full-url-in-php
+     * @author Timo Huovinen
+     *
+     * @return string
+     *   The current URL.
+     */
+    protected function getCurrentUrl()
+    {
+        $server = $this->pool->getServer();
+
+        // Check if someone has been messing with the $_SERVER, to prevent
+        // warnings and notices.
+        if (empty($server) === true ||
+            empty($server['SERVER_PROTOCOL']) === true ||
+            empty($server['SERVER_PORT']) === true ||
+            empty($server['SERVER_NAME']) === true ||
+            empty($server['REQUEST_URI']) === true
+        ) {
+            return 'n/a';
+        }
+
+        // SSL or no SSL.
+        $ssl = (!empty($server['HTTPS']) && $server['HTTPS'] === 'on');
+
+        $protocol = strtolower($server['SERVER_PROTOCOL']);
+        $protocol = substr($protocol, 0, strpos($protocol, '/'));
+        if ($ssl === true) {
+            $protocol .= 's';
+        }
+
+        $port = $server['SERVER_PORT'];
+
+        if (($ssl === false && $port === '80') || ($ssl === true && $port === '443')) {
+            // Normal combo with port and protocol.
+            $port = '';
+        } else {
+            // We have a special port here.
+            $port = ':' . $port;
+        }
+
+        if (isset($server['HTTP_HOST']) === true) {
+            $host = $server['HTTP_HOST'];
+        } else {
+            $host = $server['SERVER_NAME'] . $port;
+        }
+
+        return $this->pool->encodingService->encodeString($protocol . '://' . $host . $server['REQUEST_URI']);
     }
 }
