@@ -56,17 +56,21 @@ class File extends AbstractOutput
      *   - footer
      *   This means, that every output is split in 4 parts
      */
-    protected $chunkStrings = array();
+    protected $chunkStrings = [];
 
     /**
      * Adds output to our shutdown handler.
      *
      * @param string $chunkString
      *   The chunked output string.
+     *
+     * @return $this
+     *   For chaining.
      */
     public function addChunkString($chunkString)
     {
         $this->chunkStrings[] = $chunkString;
+        return $this;
     }
 
     /**
