@@ -34,6 +34,8 @@
 
 namespace Brainworxx\Krexx\Analyse\Callback\Analyse\Objects;
 
+use ReflectionProperty;
+
 /**
  * Analysis of protected properties.
  *
@@ -63,12 +65,12 @@ class ProtectedProperties extends AbstractObjectAnalysis
 
         /** @var \Brainworxx\Krexx\Service\Reflection\ReflectionClass $ref */
         $ref = $this->parameters[static::PARAM_REF];
-        $refProps = $ref->getProperties(\ReflectionProperty::IS_PROTECTED);
+        $refProps = $ref->getProperties(ReflectionProperty::IS_PROTECTED);
         if (empty($refProps) === true) {
             return $output;
         }
 
-        usort($refProps, array($this, 'reflectionSorting'));
+        usort($refProps, [$this, 'reflectionSorting']);
 
         return $output .
             $this->getReflectionPropertiesData(
