@@ -34,6 +34,8 @@
 
 namespace Brainworxx\Krexx\Controller;
 
+use Brainworxx\Krexx\Analyse\Model;
+
 /**
  * "Controller" for the dump (aka analysis) "action".
  *
@@ -67,16 +69,12 @@ class DumpController extends AbstractController
         // Find caller.
         $caller = $this->callerFinder->findCaller($headline, $data);
 
-        // We need to get the footer before the generating of the header,
-        // because we need to display messages in the header from the configuration.
-        $footer = $this->outputFooter($caller);
-
         // We will only allow code generation, if we were able to determine the varname.
         $this->pool->scope->setScope($caller[static::TRACE_VARNAME]);
 
         // Start the magic.
         $analysis = $this->pool->routing->analysisHub(
-            $this->pool->createClass('Brainworxx\\Krexx\\Analyse\\Model')
+            $this->pool->createClass(Model::class)
                 ->setData($data)
                 ->setName($caller[static::TRACE_VARNAME])
         );
@@ -95,10 +93,15 @@ class DumpController extends AbstractController
         // additional info, in case we are logging to a file.
         $this->pool->chunks->addMetadata($caller);
 
-        $this->outputService->addChunkString($this->outputHeader($caller[static::TRACE_TYPE]));
-        $this->outputService->addChunkString($analysis);
-        $this->outputService->addChunkString($footer);
-        $this->outputService->finalize();
+        // We need to get the footer before the generating of the header,
+        // because we need to display messages in the header from the configuration.
+        $footer = $this->outputFooter($caller);
+
+        $this->outputService
+            ->addChunkString($this->pool->render->renderHeader($caller[static::TRACE_TYPE], $this->outputCssAndJs()))
+            ->addChunkString($analysis)
+            ->addChunkString($footer)
+            ->finalize();
 
         return $this;
     }
