@@ -34,10 +34,16 @@
 
 namespace Brainworxx\Krexx;
 
-use Brainworxx\Krexx\Service\Factory\Pool;
 use Brainworxx\Krexx\Controller\AbstractController;
-use Brainworxx\Krexx\Service\Config\Fallback;
+use Brainworxx\Krexx\Controller\BacktraceController;
+use Brainworxx\Krexx\Controller\DumpController;
+use Brainworxx\Krexx\Controller\EditSettingsController;
+use Brainworxx\Krexx\Controller\ErrorController;
+use Brainworxx\Krexx\Controller\ExceptionController;
+use Brainworxx\Krexx\Controller\TimerController;
 use Brainworxx\Krexx\Service\Config\Config;
+use Brainworxx\Krexx\Service\Config\Fallback;
+use Brainworxx\Krexx\Service\Factory\Pool;
 
 /**
  * Public functions, allowing access to the kreXX debug features.
@@ -112,7 +118,7 @@ class Krexx
 
         AbstractController::$analysisInProgress = true;
 
-        static::$pool->createClass('Brainworxx\\Krexx\\Controller\\TimerController')
+        static::$pool->createClass(TimerController::class)
             ->noFatalForKrexx()
             ->timerAction($string)
             ->reFatalAfterKrexx();
@@ -139,7 +145,7 @@ class Krexx
 
         AbstractController::$analysisInProgress = true;
 
-        static::$pool->createClass('Brainworxx\\Krexx\\Controller\\TimerController')
+        static::$pool->createClass(TimerController::class)
             ->noFatalForKrexx()
             ->timerEndAction()
             ->reFatalAfterKrexx();
@@ -172,7 +178,7 @@ class Krexx
 
         AbstractController::$analysisInProgress = true;
 
-        static::$pool->createClass('Brainworxx\\Krexx\\Controller\\DumpController')
+        static::$pool->createClass(DumpController::class)
             ->noFatalForKrexx()
             ->dumpAction($data)
             ->reFatalAfterKrexx();
@@ -192,7 +198,6 @@ class Krexx
      *   An already existing backtrace.
      *
      * @api
-     *
      */
     public static function backtrace(array $backtrace = null)
     {
@@ -208,7 +213,7 @@ class Krexx
 
         AbstractController::$analysisInProgress = true;
 
-        static::$pool->createClass('Brainworxx\\Krexx\\Controller\\BacktraceController')
+        static::$pool->createClass(BacktraceController::class)
             ->noFatalForKrexx()
             ->backtraceAction($backtrace)
             ->reFatalAfterKrexx();
@@ -226,7 +231,7 @@ class Krexx
         Pool::createPool();
 
         static::$pool->config->setDisabled(true);
-        static::$pool->createClass('Brainworxx\\Krexx\\Controller\\DumpController')
+        static::$pool->createClass(DumpController::class)
             ->noFatalForKrexx();
 
         Config::$disabledByPhp = true;
@@ -251,7 +256,7 @@ class Krexx
             return;
         }
 
-         static::$pool->createClass('Brainworxx\\Krexx\\Controller\\EditSettingsController')
+         static::$pool->createClass(EditSettingsController::class)
             ->noFatalForKrexx()
             ->editSettingsAction()
             ->reFatalAfterKrexx();
@@ -261,6 +266,9 @@ class Krexx
      * Registers a shutdown function.
      *
      * Our fatal errorhandler is located there.
+     *
+     * @deprecated
+     *   Since 3.1.0. Will be removed when dropping PHP 5 support.
      *
      * @api
      */
@@ -284,7 +292,7 @@ class Krexx
             return;
         }
 
-        static::$pool->createClass('Brainworxx\\Krexx\\Controller\\ErrorController')
+        static::$pool->createClass(ErrorController::class)
             ->registerFatalAction();
     }
 
@@ -294,6 +302,9 @@ class Krexx
      * We can not unregister a once declared shutdown function,
      * so we need to tell our errorhandler to do nothing, in case
      * there is a fatal.
+     *
+     * @deprecated
+     *   Since 3.1.0. Will be removed when dropping PHP 5 support.
      *
      * @api
      */
@@ -308,8 +319,48 @@ class Krexx
             return;
         }
 
-        static::$pool->createClass('Brainworxx\\Krexx\\Controller\\ErrorController')
+        static::$pool->createClass(ErrorController::class)
             ->unregisterFatalAction();
+    }
+
+    /**
+     * Registering our exception handler.
+     *
+     * @api
+     */
+    public static function registerExceptionHandler()
+    {
+        Pool::createPool();
+
+        // Disabled?
+        if (static::$pool->config->getSetting(Fallback::SETTING_DISABLED) ||
+            Config::$disabledByPhp
+        ) {
+            return;
+        }
+
+        static::$pool->createClass(ExceptionController::class)
+            ->registerAction();
+    }
+
+    /**
+     * Ungistering our exception handler.
+     *
+     * @api
+     */
+    public static function unregisterExceptionHandler()
+    {
+        Pool::createPool();
+
+        // Disabled?
+        if (static::$pool->config->getSetting(Fallback::SETTING_DISABLED) ||
+            Config::$disabledByPhp
+        ) {
+            return;
+        }
+
+        static::$pool->createClass(ExceptionController::class)
+            ->unregisterAction();
     }
 
     /**
@@ -338,12 +389,15 @@ class Krexx
      * When there are classes found inside the backtrace,
      * they will be analysed.
      *
+     * @param array $backtrace
+     *   an already existing backtrace
+     *
      * @api
      */
-    public static function logBacktrace()
+    public static function logBacktrace(array $backtrace = null)
     {
         static::startForcedLog();
-        static::backtrace();
+        static::backtrace($backtrace);
         static::endForcedLog();
     }
 
@@ -390,6 +444,6 @@ class Krexx
     {
         // Reset everything afterwards.
         static::$pool->config = static::$pool
-            ->createClass('Brainworxx\\Krexx\\Service\\Config\\Config');
+            ->createClass(Config::class);
     }
 }
