@@ -37,285 +37,12 @@ namespace Brainworxx\Krexx\Service\Config;
 /**
  * Security measures for the configuration
  *
+ * @deprecated
+ *   Since 3.1.0. Will be removed.
+ *
  * @package Brainworxx\Krexx\Service\Config
  */
-class Security extends Fallback
+class Security extends Validation
 {
 
-    const KEY_CONFIG_ERROR = 'configError';
-    const KEY_CONFIG_ERROR_BOOL = 'configErrorBool';
-    const KEY_CONFIG_ERROR_INT = 'configErrorInt';
-    const KEY_CONFIG_ERROR_DEBUG_BLACKLIST = 'configErrorDebugBlacklist';
-    const KEY_CONFIG_ERROR_DEBUG_INVALID = 'configErrorDebugInvalid';
-
-    /**
-     * Setter for the debug method blacklist by the config class.
-     *
-     * @param array $methodBlacklist
-     *   The blacklist.
-     * @return $this
-     *   This, for chaining.
-     */
-    public function setMethodBlacklist(array $methodBlacklist)
-    {
-        $this->methodBlacklist = $methodBlacklist;
-        return $this;
-    }
-
-    /**
-     * Evaluate a single setting from the cookies or the ini file.
-     *
-     * @param string $group
-     *   The group value in the ini.
-     * @param string $name
-     *   The name of the setting.
-     * @param string $value
-     *   The value to evaluate.
-     *
-     * @return bool
-     *   If it was evaluated.
-     */
-    public function evaluateSetting($group, $name, $value)
-    {
-        if ($group === 'feEditing') {
-            // Logging options can never be changed in the frontend.
-            // The debug methods will also not be editable.
-            return !in_array($name, $this->feConfigNoEdit);
-        }
-
-        // We simply call the configured evaluation method.
-        $callback = $this->feConfigFallback[$name][static::EVALUATE];
-        return $this->$callback($value, $name, $group);
-    }
-
-    /**
-     * We check the dev handle for a-z and A-Z.
-     *
-     * @param string $value
-     *   The value we want to evaluate
-     * @param string $name
-     *   The name of the value we are checking, needed for the feedback text.
-     *
-     * @return boolean
-     *   Whether it does evaluate or not.
-     */
-    protected function evalDevHandle($value, $name)
-    {
-        $result = preg_match('/[^a-zA-Z]/', $value) === 0;
-        if ($result === false) {
-            $this->pool->messages->addMessage(static::KEY_CONFIG_ERROR . ucfirst($name));
-        }
-
-        return $result;
-    }
-
-    /**
-     * We check the configuration for this skin.
-     *
-     * @param string $value
-     *   The value we want to evaluate
-     * @param string $name
-     *   The name of the value we are checking, needed for the feedback text.
-     *
-     * @return boolean
-     *   Whether it does evaluate or not.
-     */
-    protected function evalSkin($value, $name)
-    {
-        $result = isset($this->skinConfiguration[$value]) &&
-            class_exists($this->skinConfiguration[$value][static::SKIN_CLASS]) &&
-            $this->pool->fileService->fileIsReadable(
-                $this->skinConfiguration[$value][static::SKIN_DIRECTORY] . 'header.html'
-            );
-
-        if ($result === false) {
-            $this->pool->messages->addMessage(static::KEY_CONFIG_ERROR . ucfirst($name));
-        }
-
-        return $result;
-    }
-
-    /**
-     * We are expecting 'browser' or 'file'.
-     *
-     * @param string $value
-     *   The value we want to evaluate
-     * @param string $name
-     *   The name of the value we are checking, needed for the feedback text.
-     *
-     * @return boolean
-     *   Whether it does evaluate or not.
-     */
-    protected function evalDestination($value, $name)
-    {
-        $result = ($value === static::VALUE_BROWSER || $value === 'file');
-        if ($result === false) {
-            $this->pool->messages->addMessage(static::KEY_CONFIG_ERROR . ucfirst($name));
-        }
-
-        return $result;
-    }
-
-    /**
-     * Evaluating the IP range, by testing that it is not empty.
-     *
-     * @param string $value
-     *   The value we want to evaluate
-     * @param string $name
-     *   The name of the value we are checking, needed for the feedback text.
-     *
-     * @return boolean
-     *   Whether it does evaluate or not.
-     */
-    protected function evalIpRange($value, $name)
-    {
-        $result = empty($value);
-        if ($result === true) {
-            $this->pool->messages->addMessage(static::KEY_CONFIG_ERROR . ucfirst($name));
-        }
-
-        return !$result;
-    }
-
-    /**
-     * Evaluation the maximum runtime, by looking at the server settings, as
-     * well as checking for an integer value.
-     *
-     * @param string $value
-     *   The value we want to evaluate
-     * @param string $name
-     *   The name of the value we are checking, needed for the feedback text.
-     * @param string $group
-     *   The name of the group that we are evaluating, needed for the feedback
-     *   text.
-     *
-     * @return boolean
-     *   Whether it does evaluate or not.
-     */
-    protected function evalMaxRuntime($value, $name, $group)
-    {
-        // Check for integer first.
-        if ($this->evalInt($value, $name, $group) === false) {
-            return false;
-        }
-
-        $maxTime = (int)ini_get('max_execution_time');
-        // We need a maximum runtime in the first place
-        // and then check, if we have a value smaller than it.
-        if ($maxTime <= 0) {
-            // We were unable to get the maximum runtime from the server.
-            // No need to check any further.
-            return true;
-        }
-
-        if ($maxTime < (int)$value) {
-            $this->pool->messages->addMessage(
-                static::KEY_CONFIG_ERROR . ucfirst($name) . 'Big',
-                array($maxTime)
-            );
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * Evaluates a string of 'true' or 'false'.
-     *
-     * @param string $value
-     *   The string we want to evaluate.
-     * @param string $name
-     *   The name of the value we are checking, needed for the feedback text.
-     * @param string $group
-     *   The name of the group that we are evaluating, needed for the feedback
-     *   text.
-     *
-     * @return bool
-     *   Whether it does evaluate or not.
-     */
-    protected function evalBool($value, $name, $group)
-    {
-        $result = ($value === static::VALUE_TRUE || $value === static::VALUE_FALSE);
-        if ($result === false) {
-            $this->pool->messages->addMessage(static::KEY_CONFIG_ERROR_BOOL, array($group, $name));
-        }
-
-        return $result;
-    }
-
-    /**
-     * Evaluates a string of integer.
-     *
-     * It must be greater than 0 and smaller than 101.
-     *
-     * @param string $value
-     *   The string we want to evaluate.
-     * @param string $name
-     *   The name of the value we are checking, needed for the feedback text.
-     * @param string $group
-     *   The name of the group that we are evaluating, needed for the feedback
-     *   text.
-     *
-     * @return bool
-     *   Whether it does evaluate or not.
-     */
-    protected function evalInt($value, $name, $group)
-    {
-        $result = ((int) $value) > 0;
-        if ($result === false) {
-            $this->pool->messages->addMessage(static::KEY_CONFIG_ERROR_INT, array($group, $name));
-        }
-
-        return $result;
-    }
-
-    /**
-     * We do not evaluate this one.
-     *
-     * @return bool
-     *   Always true, we do not eval this one.
-     */
-    protected function doNotEval()
-    {
-        return true;
-    }
-
-    /**
-     * Sanity check, if the supplied debug methods are not obviously flawed.
-     *
-     * @param string $value
-     *   Comma separated list of debug methods.
-     * @param string $name
-     *   The name of the value we are checking, needed for the feedback text.
-     * @param string $group
-     *   The name of the group that we are evaluating, needed for the feedback
-     *   text.
-     *
-     * @return bool
-     *   Whether it does evaluate or not.
-     */
-    protected function evalDebugMethods($value, $name, $group)
-    {
-        $list = explode(',', $value);
-
-        foreach ($list as $entry) {
-            // Test for whitespace and the blacklist.
-            if (in_array($entry, $this->methodBlacklist)) {
-                $this->pool->messages->addMessage(
-                    static::KEY_CONFIG_ERROR_DEBUG_BLACKLIST,
-                    array($group, $name, $entry)
-                );
-                return false;
-            }
-            if (strpos($entry, ' ') !== false) {
-                $this->pool->messages->addMessage(
-                    static::KEY_CONFIG_ERROR_DEBUG_INVALID,
-                    array($group, $name, $entry)
-                );
-                return false;
-            }
-        }
-
-        return true;
-    }
 }
