@@ -34,6 +34,8 @@
 
 namespace Brainworxx\Krexx\Service\Reflection;
 
+use ReflectionProperty;
+
 /**
  * The original \ReflectionProperty may throw an error when used with
  * dynamically declared properties
@@ -52,7 +54,7 @@ namespace Brainworxx\Krexx\Service\Reflection;
  *
  * @package Brainworxx\Krexx\Service\Misc
  */
-class UndeclaredProperty extends \ReflectionProperty
+class UndeclaredProperty extends ReflectionProperty
 {
     /**
      * This one is always undeclared.
